Ground Pound, Walking Spider Initial AI, Sandbox Level, Basic Damage/Life Implementation, Basic Spider Stun
Basic Walking Spider AI that paces back and forth, and damages the player when it collides with the spider's mouth. The player can either land on the spider with a regular or double jump for a minor stun, or can use the new simple ground pound (jump then down) to damage and destroy the spiders. Also includes a new Sandbox level for development and testing.